Updating composer bootstrapping to work with existing sites that use composer.

+++ core/bootstrap.php
@@ -33,24 +33,61 @@
 		include SERVER_ROOT."core/inc/bigtree/utils.php";
 	}
 
-	// Include Composer's autoloader
-	if (!file_exists(SERVER_ROOT."vendor/autoload.php")) {
-		BigTree::makeDirectory(SERVER_ROOT."vendor/");
+	// Setup the vendor folder if we haven't done this check before
+	if (!file_exists(SERVER_ROOT."cache/composer-check.flag")) {
+		if (!file_exists(SERVER_ROOT."vendor/")) {
+			BigTree::makeDirectory(SERVER_ROOT."vendor/");
+		}
 
 		$path = str_replace("core/bootstrap.php", "", __FILE__);
 		$off_path = str_replace(SERVER_ROOT, "", $path);
 
-		symlink(SERVER_ROOT.$off_path."vendor/autoload.php", SERVER_ROOT."vendor/autoload.php");
-		symlink(SERVER_ROOT.$off_path."vendor/aws", SERVER_ROOT."vendor/aws");
-		symlink(SERVER_ROOT.$off_path."vendor/composer", SERVER_ROOT."vendor/composer");
-		symlink(SERVER_ROOT.$off_path."vendor/guzzlehttp", SERVER_ROOT."vendor/guzzlehttp");
-		symlink(SERVER_ROOT.$off_path."vendor/mtdowling", SERVER_ROOT."vendor/mtdowling");
-		symlink(SERVER_ROOT.$off_path."vendor/oyejorge", SERVER_ROOT."vendor/oyejorge");
-		symlink(SERVER_ROOT.$off_path."vendor/psr", SERVER_ROOT."vendor/psr");
+		if (!file_exists(SERVER_ROOT."vendor/autoload.php")) {
+			BigTree::copyFile(SERVER_ROOT.$off_path."vendor/autoload.php", SERVER_ROOT."vendor/autoload.php");
+		}
 
-		BigTree::copyFile(SERVER_ROOT.$off_path."vendor/autoload.php", SERVER_ROOT."vendor/autoload.php");
-		BigTree::copyFile(SERVER_ROOT.$off_path."composer.lock", SERVER_ROOT."composer.lock");
-		BigTree::copyFile(SERVER_ROOT.$off_path."composer.json", SERVER_ROOT."composer.json");
+		if (!file_exists(SERVER_ROOT."vendor/aws/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/aws", SERVER_ROOT."vendor/aws");
+		}
+
+		if (!file_exists(SERVER_ROOT."vendor/composer/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/composer", SERVER_ROOT."vendor/composer");
+		}
+
+		if (!file_exists(SERVER_ROOT."vendor/guzzlehttp/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/guzzlehttp", SERVER_ROOT."vendor/guzzlehttp");
+		}
+
+		if (!file_exists(SERVER_ROOT."vendor/mtdowling/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/mtdowling", SERVER_ROOT."vendor/mtdowling");
+		}
+
+		if (!file_exists(SERVER_ROOT."vendor/oyejorge/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/oyejorge", SERVER_ROOT."vendor/oyejorge");
+		}
+
+		if (!file_exists(SERVER_ROOT."vendor/psr/")) {
+			symlink(SERVER_ROOT.$off_path."vendor/psr", SERVER_ROOT."vendor/psr");
+		}
+
+		if (file_exists(SERVER_ROOT."composer.json") && !is_writable(SERVER_ROOT."composer.json")) {
+			die("BigTree needs to write to your composer.json file to ensure a composer update does not wipe required libraries.");
+		}
+
+		if (file_exists(SERVER_ROOT."composer.json")) {
+			$bigtree_composer_json = json_decode(SERVER_ROOT.$off_path."composer.json", true);
+			$existing_json = json_decode(SERVER_ROOT."composer.json", true);
+
+			foreach ($bigtree_composer_json["require"] as $key => $value) {
+				if (!isset($existing_json["require"][$key])) {
+					$existing_json["require"][$key] = $value;
+				}
+			}
+
+			BigTree::putFile(SERVER_ROOT."composer.json", BigTree::json($existing_json));
+		}
+
+		BigTree::putFile(SERVER_ROOT."cache/composer-check.flag", "done");
 	}
 
 	include SERVER_ROOT."vendor/autoload.php";
